refactor: change css to be used without ClassNames component
